Fix event routing, add config files for env variables

+++ media-service/src/eventHandlers/media-event-handlers.js
@@ -3,25 +3,36 @@ const { deleteMediaFromCloudinary } = require("../utils/cloudinary");
 const logger = require("../utils/logger");
 
 const handlePostDeleted = async (event) => {
-  // console.log({ "post-delete-event": event });
+  logger.info('Received event:', { event });
+
   const { postId, mediaIds } = event;
+
+  // Ensure event has the required fields
+  if (!postId || !mediaIds || !Array.isArray(mediaIds)) {
+    logger.warn('Invalid or no post.deleted event received:', { event });
+    return;
+  }
 
   try {
     // Delete media associated with the post
     const mediaToDelete = await Media.find({ _id: { $in: mediaIds } });
 
-    // Delete media from cloudinary
-    mediaToDelete.forEach(async (media) => {
+    // Delete media from Cloudinary
+    for (const media of mediaToDelete) {
       await deleteMediaFromCloudinary(media.publicId);
       await Media.findByIdAndDelete(media._id);
-
-      logger.info(`Deleted media ${media._id} associated with deleted post ${postId}`);
-    });
+      logger.info(
+        `Deleted media ${media._id} associated with deleted post ${postId}`
+      );
+    }
 
     logger.info(`Processed deletion of media for post id: ${postId}`);
   } catch (error) {
-    logger.error("Error handling media deletion with post deleted event", error);
+    logger.error(
+      'Error handling media deletion with post.deleted event',
+      error
+    );
   }
-}
+};
 
 module.exports = { handlePostDeleted };
